Rename term_command to terminal_command
