+++ crates/assistant/src/slash_command/terminal_command.rs
